refactor(interactions): migrate application buttons + modals to registry
application_vote keeps its underscore-prefix customId (persisted in
live voting embeds). The submit-error officer alert path (#42) is
preserved inside the confirm handler's own try/catch.
